Fix gcc compiler warning treated as error: sign-compare
Summary: Updating D33

Test Plan: -

Reviewers: O1 MITK Core, goch

Reviewed By: O1 MITK Core, goch

Subscribers: kalali, goch

Maniphest Tasks: T19935

Differential Revision: https://phabricator.mitk.org/D34

+++ Modules/Core/src/Algorithms/mitkExtractSliceFilter.cpp
@@ -377,7 +377,7 @@ void mitk::ExtractSliceFilter::GenerateData()
 
     /*================ #BEGIN Extract component from image slice ================*/
     int numberOfScalarComponent = reslicedImage->GetNumberOfScalarComponents();
-    if (numberOfScalarComponent > 1 && numberOfScalarComponent >= m_Component)
+    if (numberOfScalarComponent > 1 && static_cast<unsigned int>(numberOfScalarComponent) >= m_Component)
     {
       // image has more than one component, extract the correct component information with the given 'component' parameter
       auto vectorComponentExtractor = vtkSmartPointer<vtkImageExtractComponents>::New();
